Updated module augmentation for paragraph package.
