+++ packages/ckeditor5-paragraph/src/augmentation.ts
@@ -1,0 +1,19 @@
+/**
+ * @license Copyright (c) 2003-2023, CKSource Holding sp. z o.o. All rights reserved.
+ * For licensing, see LICENSE.md or https://ckeditor.com/legal/ckeditor-oss-license
+ */
+
+import type InsertParagraphCommand from './insertparagraphcommand';
+import type Paragraph from './paragraph';
+import type ParagraphCommand from './paragraphcommand';
+
+declare module '@ckeditor/ckeditor5-core' {
+	interface CommandsMap {
+		insertParagraph: InsertParagraphCommand;
+		paragraph: ParagraphCommand;
+	}
+
+	interface PluginsMap {
+		[ Paragraph.pluginName ]: Paragraph;
+	}
+}
