+++ packages/ckeditor5-paragraph/src/index.ts
@@ -10,3 +10,5 @@
 export { default as Paragraph } from './paragraph';
 export { default as ParagraphButtonUI } from './paragraphbuttonui';
 export type { default as ParagraphCommand } from './paragraphcommand';
+
+import './augmentation';
