+++ packages/ckeditor5-paragraph/src/insertparagraphcommand.ts
@@ -70,9 +70,3 @@ export default class InsertParagraphCommand extends Command {
 		} );
 	}
 }
-
-declare module '@ckeditor/ckeditor5-core' {
-	interface CommandsMap {
-		insertParagraph: InsertParagraphCommand;
-	}
-}
